did some huge changes in dclient.c
I've putted a struct in the library, this is what the server and the client will be communicating!

( we have to gather in order to make some choices about the project! )

TODO: see the rest of the communication in the dserver.c (didnt touched it yet...)

TODO: to strcpy on the lines 47-52 because yes ( i need to check if it is important to do that )

TODO: see about consistensy ( i bad english (.)(.) ) we have smth like that in guide 1

TODO: do the rest the functions (at least the communication bit)

im going to sleep rn goodbye:!:::!::