fixed the server problem
The issue is here:
read(fd_to_server, &fileinfo, sizeof(FileInfo));

We are not checking the return value of read().
If the client finishes and closes the fifo, the server's read will return 0, meaning EOF (end-of-file), and then the server will continue in the loop but the fileinfo will be invalid That’s why the server dies and nothing happens, the server is stuck with a closed fifo and not reading new client requests!

the new things are from line 61 till line 89

and put the H_library